Can calculate, but has bugs
Made the calculator functional, refactoring much of the code in the process. It has bugs that need working out, particularly when working with multiple operations without explicitly clicking "=".  Utilizes a recursive function for evaluation that is capable of following PEMDAS.

P.S. - forgot to mention but it now has an expressions section, which now stores prior expressions. It occurred to me that this is essentially a history, but still, I thought it was a neat addition.