Base PEMDAS calculator complete!
So silly me. I used "eval" as a function name originally, and recalling this was an innate js function, I renamed it. I failed, however, to rename the recursive calls to it. The lack of an undefined function error caused the problem to elude me. On the bright side, after some tweaks, it works! All that's left is to make the decimal button I added functional.